+++ src/main/scala/spoker/betting/stack/StackManagement.scala
@@ -3,7 +3,7 @@ package spoker.betting.stack
 import scala.collection.mutable.Map
 import spoker.internals.MutableStateManagement
 
-trait StackManagement extends MutableStateManagement[Int, Int] {
-  override def updateState(holderName: String, su: Int): Int =
+trait StackManagement extends MutableStateManagement[Double, Double] {
+  override def updateState(holderName: String, su: Double): Double =
     currentState(holderName) + su
 }
